+++ interfaces/yearn/IVault.sol
@@ -4,15 +4,18 @@ pragma solidity ^0.6.8;
 
 interface IVault {
     function token() external view returns (address);
+
     function underlying() external view returns (address);
 
     function name() external view returns (string memory);
+
     function symbol() external view returns (string memory);
+
     function decimals() external view returns (uint8);
-    
+
     function controller() external view returns (address);
+
     function governance() external view returns (address);
 
-    function getPricePerFullShare() external view returns (uint);
-    
+    function getPricePerFullShare() external view returns (uint256);
 }
