+++ interfaces/yearn/IWrappedVault.sol
@@ -6,13 +6,14 @@ interface IWrappedVault {
     function token() external view returns (address);
 
     function name() external view returns (string memory);
+
     function symbol() external view returns (string memory);
+
     function decimals() external view returns (uint8);
-    
+
     function governance() external view returns (address);
-    
+
     function vault() external view returns (address);
 
-    function getPricePerFullShare() external view returns (uint);
-    
+    function getPricePerFullShare() external view returns (uint256);
 }
